+++ TP8_MVC/Controllers/ProductosController.cs
@@ -1,22 +1,35 @@
 using Microsoft.AspNetCore.Mvc;
+using TP8_MVC.Interfaces;
 using TP8_MVC.Models;
-using TP8_MVC.Repositories;
 using TP8_MVC.ViewModels;
 
 namespace TP8_MVC.Controllers
 {
     public class ProductosController : Controller
     {
-        private readonly ProductoRepository _productoRepository;
+        private readonly IProductoRepository _productoRepository;
+        private readonly IAuthenticationService _authenticationService;
 
-        public ProductosController()
+        public ProductosController(IProductoRepository productoRepository, IAuthenticationService authenticationService)
         {
-            _productoRepository = new ProductoRepository();
+            _productoRepository = productoRepository;
+            _authenticationService = authenticationService;
         }
 
         // GET: Productos
         public IActionResult Index()
         {
+            // Solo administradores pueden gestionar productos
+            if (!_authenticationService.IsAuthenticated())
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
+            if (!_authenticationService.HasAccessLevel("Administrador"))
+            {
+                return RedirectToAction("AccesoDenegado", "Home");
+            }
+
             var productos = _productoRepository.GetAll();
             return View(productos);
         }
@@ -24,6 +37,16 @@ namespace TP8_MVC.Controllers
         // GET: Productos/Create
         public IActionResult Create()
         {
+            if (!_authenticationService.IsAuthenticated())
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
+            if (!_authenticationService.HasAccessLevel("Administrador"))
+            {
+                return RedirectToAction("AccesoDenegado", "Home");
+            }
+
             return View(new ProductoViewModel());
         }
 
@@ -32,6 +55,16 @@ namespace TP8_MVC.Controllers
         [ValidateAntiForgeryToken]
         public IActionResult Create(ProductoViewModel productoVM)
         {
+            if (!_authenticationService.IsAuthenticated())
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
+            if (!_authenticationService.HasAccessLevel("Administrador"))
+            {
+                return RedirectToAction("AccesoDenegado", "Home");
+            }
+
             // 1. Chequeo de seguridad del servidor
             if (!ModelState.IsValid)
             {
@@ -55,6 +88,16 @@ namespace TP8_MVC.Controllers
         // GET: Productos/Edit/5
         public IActionResult Edit(int id)
         {
+            if (!_authenticationService.IsAuthenticated())
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
+            if (!_authenticationService.HasAccessLevel("Administrador"))
+            {
+                return RedirectToAction("AccesoDenegado", "Home");
+            }
+
             var producto = _productoRepository.GetById(id);
             if (producto == null)
             {
@@ -77,6 +120,16 @@ namespace TP8_MVC.Controllers
         [ValidateAntiForgeryToken]
         public IActionResult Edit(int id, ProductoViewModel productoVM)
         {
+            if (!_authenticationService.IsAuthenticated())
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
+            if (!_authenticationService.HasAccessLevel("Administrador"))
+            {
+                return RedirectToAction("AccesoDenegado", "Home");
+            }
+
             if (id != productoVM.IdProducto)
             {
                 return NotFound();
@@ -105,6 +158,16 @@ namespace TP8_MVC.Controllers
         // GET: Productos/Delete/5
         public IActionResult Delete(int id)
         {
+            if (!_authenticationService.IsAuthenticated())
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
+            if (!_authenticationService.HasAccessLevel("Administrador"))
+            {
+                return RedirectToAction("AccesoDenegado", "Home");
+            }
+
             var producto = _productoRepository.GetById(id);
             if (producto == null)
             {
@@ -118,6 +181,16 @@ namespace TP8_MVC.Controllers
         [ValidateAntiForgeryToken]
         public IActionResult DeleteConfirmed(int id)
         {
+            if (!_authenticationService.IsAuthenticated())
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
+            if (!_authenticationService.HasAccessLevel("Administrador"))
+            {
+                return RedirectToAction("AccesoDenegado", "Home");
+            }
+
             _productoRepository.Delete(id);
             return RedirectToAction(nameof(Index));
         }
